requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
django-cors-headers
requests
pyyaml
djangorestframework-simplejwt
Pillow
dj-database-url
gunicorn
python-decouple
psycopg[binary]
psycopg2-binary
